fix: move css utils from utils to css
