+++ packages/builder/builder-shared/src/utils.ts
@@ -1,13 +1,7 @@
-import {
-  CSS_MODULES_REGEX,
-  GLOBAL_CSS_REGEX,
-  NODE_MODULES_REGEX,
-} from './constants';
 import type {
   SharedNormalizedConfig,
   BuilderTarget,
   SharedCompiledPkgNames,
-  CssModules,
 } from './types';
 import { join } from 'path';
 
@@ -18,14 +12,6 @@ export const removeLeadingSlash = (s: string): string => s.replace(/^\/+/, '');
 
 export const addTrailingSlash = (s: string): string =>
   s.endsWith('/') ? s : `${s}/`;
-
-/** Determine if a file path is a CSS module when disableCssModuleExtension is enabled. */
-export const isLooseCssModules = (path: string) => {
-  if (NODE_MODULES_REGEX.test(path)) {
-    return CSS_MODULES_REGEX.test(path);
-  }
-  return !GLOBAL_CSS_REGEX.test(path);
-};
 
 export interface AwaitableGetter<T> extends PromiseLike<T[]> {
   promises: Promise<T>[];
@@ -85,43 +71,3 @@ export function isWebTarget(target: BuilderTarget | BuilderTarget[]) {
     (Array.isArray(target) ? target : [target]).includes(t as BuilderTarget),
   );
 }
-
-export type CssLoaderModules =
-  | boolean
-  | string
-  | {
-      auto: boolean | RegExp | ((filename: string) => boolean);
-    };
-
-export const isCssModules = (filename: string, modules: CssLoaderModules) => {
-  if (typeof modules === 'boolean') {
-    return modules;
-  }
-
-  // todo: this configuration is not common and more complex.
-  if (typeof modules === 'string') {
-    return true;
-  }
-
-  const { auto } = modules;
-
-  if (typeof auto === 'boolean') {
-    return auto && CSS_MODULES_REGEX.test(filename);
-  } else if (auto instanceof RegExp) {
-    return auto.test(filename);
-  } else if (typeof auto === 'function') {
-    return auto(filename);
-  }
-  return true;
-};
-
-export const getCssModulesAutoRule = (
-  config?: CssModules,
-  disableCssModuleExtension = false,
-) => {
-  if (!config || config?.auto === undefined) {
-    return disableCssModuleExtension ? isLooseCssModules : true;
-  }
-
-  return config.auto;
-};
