+++ packages/builder/builder-shared/tests/css.test.ts
@@ -1,5 +1,10 @@
 import { expect, describe, it } from 'vitest';
-import { normalizeCssLoaderOptions } from '../src/css';
+import {
+  normalizeCssLoaderOptions,
+  isCssModules,
+  getCssModulesAutoRule,
+  isLooseCssModules,
+} from '../src/css';
 
 describe('normalizeCssLoaderOptions', () => {
   it('should enable exportOnlyLocals correctly', () => {
@@ -33,4 +38,76 @@ describe('normalizeCssLoaderOptions', () => {
       },
     });
   });
+});
+
+it('check isCssModules', () => {
+  expect(isCssModules('src/index.css', false)).toBeFalsy();
+  expect(isCssModules('src/index.css', { auto: false })).toBeFalsy();
+  expect(isCssModules('src/index.module.css', { auto: false })).toBeFalsy();
+
+  expect(isCssModules('src/index.css', true)).toBeTruthy();
+
+  expect(isCssModules('src/index.css', { auto: true })).toBeFalsy();
+  expect(isCssModules('src/index.module.css', { auto: true })).toBeTruthy();
+
+  expect(
+    isCssModules('src/index.module.css', {
+      auto: path => {
+        return path.includes('.module.');
+      },
+    }),
+  ).toBeTruthy();
+
+  expect(
+    isCssModules('src/index.css', {
+      auto: path => {
+        return path.includes('.module.');
+      },
+    }),
+  ).toBeFalsy();
+
+  expect(
+    isCssModules('src/index.module.css', {
+      auto: /\.module\./i,
+    }),
+  ).toBeTruthy();
+
+  expect(
+    isCssModules('src/index.css', {
+      auto: /\.module\./i,
+    }),
+  ).toBeFalsy();
+});
+
+it('getCssModulesAutoRule', () => {
+  expect(getCssModulesAutoRule(undefined, true)).toEqual(isLooseCssModules);
+
+  expect(
+    getCssModulesAutoRule(
+      {
+        auto: false,
+      },
+      true,
+    ),
+  ).toBeFalsy();
+
+  expect(
+    getCssModulesAutoRule(
+      {
+        auto: true,
+      },
+      true,
+    ),
+  ).toBeTruthy();
+
+  const autoFn = () => true;
+
+  expect(
+    getCssModulesAutoRule(
+      {
+        auto: autoFn,
+      },
+      true,
+    ),
+  ).toEqual(autoFn);
 });
